requirements.txt
grpcio
numpy
tensorflow
vald-client-python
